Create polynomials from Sympy expressions

+++ tests/models/test_ising_model.py
@@ -5,14 +5,16 @@ from src.simulated_bifurcation.models import Ising
 
 def test_ising():
     torch.manual_seed(42)
-    J = torch.tensor([[0.0, -2.0, 3.0], [-2.0, 0.0, 1.0], [3.0, 1.0, 0.0]])
-    h = torch.tensor([1.0, -4.0, 2.0])
-    model = Ising(J, h, device=torch.device("cpu"))
+    J = torch.tensor([[0, -2, 3], [-2, 0, 1], [3, 1, 0]])
+    h = torch.tensor([1, -4, 2])
+    model = Ising(J, h, dtype=torch.float32, device=torch.device("cpu"))
     spin_vector, value = model.optimize(
         agents=10,
         verbose=False,
         best_only=True,
         ballistic=True,
     )
-    assert torch.equal(torch.tensor([-1.0, 1.0, -1.0]), spin_vector)
+    assert torch.equal(
+        torch.tensor([-1.0, 1.0, -1.0], dtype=torch.float32), spin_vector
+    )
     assert -11.0 == value
